# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para la gestión de estudiantes, materias y calificaciones",
        "title": "API de Control Escolar",
        "contact": {
            "name": "Soporte API",
            "email": "estefany.montiel@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api",
    "paths": {
        "/grades": {
            "post": {
                "description": "Registra una nueva calificación para un estudiante en una materia",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Crear una nueva calificación",
                "parameters": [
                    {
                        "description": "Información de la calificación",
                        "name": "grade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Grade"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Grade"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/grades/student/{student_id}": {
            "get": {
                "description": "Obtiene todas las calificaciones registradas para un estudiante específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Obtener todas las calificaciones de un estudiante",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del estudiante",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Grade"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/grades/{grade_id}": {
            "put": {
                "description": "Actualiza el valor de una calificación existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Actualizar una calificación",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la calificación",
                        "name": "grade_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nueva información de la calificación",
                        "name": "grade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Grade"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Grade"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina una calificación del sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Eliminar una calificación",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la calificación",
                        "name": "grade_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/grades/{grade_id}/student/{student_id}": {
            "get": {
                "description": "Obtiene una calificación específica de un estudiante por grade_id y student_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Obtener calificación específica",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la calificación",
                        "name": "grade_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del estudiante",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Grade"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Grade": {
            "type": "object",
            "required": [
                "grade",
                "student_id",
                "subject_id"
            ],
            "properties": {
                "grade": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 95.5
                },
                "grade_id": {
                    "type": "integer",
                    "example": 1
                },
                "student": {
                    "$ref": "#/definitions/models.Student"
                },
                "student_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "subject": {
                    "$ref": "#/definitions/models.Subject"
                },
                "subject_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                }
            }
        },
        "models.Student": {
            "type": "object",
            "required": [
                "email",
                "group",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "maria.garcia@escuela.com"
                },
                "group": {
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 1,
                    "example": "5A"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "María García"
                },
                "student_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.Subject": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Matemáticas"
                },
                "subject_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}